basic code

+++ MinesweeperBackend/IMineSweeperPlayer.cs
@@ -4,6 +4,11 @@ namespace MinesweeperBackend
 {
 	public interface IMineSweeperPlayer : IGenericPlayer
 	{
-		public (int, int) MakeMove(int?[,] boardState);
+		/// <summary>
+		/// 
+		/// </summary>
+		/// <param name="boardState">Null if not uncovered, otherwise the number of surrounding bombs.</param>
+		/// <returns></returns>
+		public (int, int) MakeMove(byte?[,] boardState);
 	}
 }
